requirements.txt
inky[impression]==1.4.0
requests==2.31.0
Pillow==10.0.1
schedule==1.2.0
python-dotenv==1.0.0
